setup_firebase_functions: adjusted functions and deployed to dev environment
