+++ .env.template
@@ -1,10 +1,5 @@
 ENVIRONMENT=
 JWT_SECRET=
-FB_API_KEY=
-FB_AUTH_DOMAIN=
 FB_PROJECT_ID=
-FB_STORAGE_BUCKET=
-FB_MESSAGING_SENDER_ID=
-FB_APP_ID=
 FB_CLIENT_EMAIL=
 FB_PRIVATE_KEY=
