+++ package.json
@@ -8,18 +8,26 @@
   "scripts": {
     "build": "nest build",
     "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
-    "start": "nest start",
-    "start:dev": "npm run set-environment -- development && nest start --watch",
-    "start:debug": "nest start --debug --watch",
-    "start:prod": "node dist/main",
+    "start": "npm run use-default && npm run build && firebase emulators:start --only functions",
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
     "test": "jest",
     "test:watch": "jest --watch",
     "test:cov": "jest --coverage",
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
     "test:e2e": "jest --config ./test/jest-e2e.json",
-    "set-environment": "node scripts/switch-environment.mjs"
+    "set-environment": "node scripts/switch-environment.mjs",
+    "shell": "npm run build && firebase functions:shell",
+    "deploy": "npm run use-default && firebase deploy --only functions",
+    "deploy:prod": "npm run use-prod && firebase deploy --only functions",
+    "logs": "firebase functions:log",
+    "firebase-use": "firebase use $PROJECT_ALIAS",
+    "use-default": "PROJECT_ALIAS=default npm run firebase-use",
+    "use-prod": "PROJECT_ALIAS=prod npm run firebase-use"
   },
+  "engines": {
+    "node": "18"
+  },
+  "main": "dist/index.js",
   "dependencies": {
     "@nestjs/common": "^10.0.0",
     "@nestjs/config": "^3.0.0",
@@ -31,6 +39,7 @@
     "class-validator": "^0.14.0",
     "firebase": "^10.0.0",
     "firebase-admin": "^11.9.0",
+    "firebase-functions": "^4.3.1",
     "nestjs-pino": "^3.3.0",
     "pino-http": "^8.3.3",
     "pino-pretty": "^10.0.0",
@@ -60,7 +69,8 @@
     "ts-loader": "^9.4.3",
     "ts-node": "^10.9.1",
     "tsconfig-paths": "^4.2.0",
-    "typescript": "^5.1.3"
+    "typescript": "^5.1.3",
+    "firebase-functions-test": "^3.1.0"
   },
   "jest": {
     "moduleFileExtensions": [
