+++ src/environments/environment.interface.ts
@@ -1,10 +1,5 @@
-import { FirebaseOptions } from '@firebase/app';
-import { ServiceAccount } from 'firebase-admin/lib/app/credential';
-
 export interface Environment {
   port: number;
   apiUrl: string;
   jwtSecret: string;
-  firebaseConfig: FirebaseOptions;
-  serviceAccount: ServiceAccount;
 }
