+++ src/modules/firebase/firestore/index.ts
@@ -1,3 +1,2 @@
-export * from './firestore.service';
 export * from './firestore.base';
 export * from './firestore.enum';
