+++ src/modules/owners/owners.service.ts
@@ -1,12 +1,8 @@
 import { Injectable } from '@nestjs/common';
-import { FirestoreBase, FirestoreService } from '../firebase/firestore';
+import { FirestoreBase } from '../firebase/firestore';
 import { Owner } from './owner.schema';
 
 @Injectable()
 export class OwnersService extends FirestoreBase<Owner> {
   protected readonly collectionName: string = 'owners';
-
-  constructor(firestoreService: FirestoreService) {
-    super(firestoreService);
-  }
 }
