+++ src/modules/vehicles/vehicles.service.ts
@@ -1,11 +1,8 @@
 import { Injectable } from '@nestjs/common';
-import { FirestoreBase, FirestoreService } from '../firebase/firestore';
+import { FirestoreBase } from '../firebase/firestore';
 import { Vehicle } from './vehicle.schema';
 
 @Injectable()
 export class VehiclesService extends FirestoreBase<Vehicle> {
   protected readonly collectionName: string = 'vehicles';
-  constructor(firestoreService: FirestoreService) {
-    super(firestoreService);
-  }
 }
